+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - home volume is set to the namespace name if a namespace is defined.
 - allow passing args directly to a nix run command: `podenv nix:github:podenv/devenv#emacs --version`
 - fix syscaps argument for bwrap.
+- use system chcon to set label.
 
 ## 0.1.0 (2022-05-02)