+++ CHANGELOG.md
@@ -2,14 +2,14 @@
 
 ## 0.3.0 (next)
 
-## 0.2.0 (2022-07-10)
+## 0.2.0 (2022-07-26)
 
 - restrict ssh capability to the agent socket.
 - home volume is set to the namespace name if a namespace is defined.
-- allow passing args directly to a nix run command: `podenv nix:github:podenv/devenv#emacs --version`
+- allow passing args directly to a command: `podenv nix:github:podenv/devenv#emacs --version`
 - fix syscaps argument for bwrap.
 - use system chcon to set label.
-- add --detach argument.
+- add --manifest argument to show every configuration context.
 
 ## 0.1.0 (2022-05-02)